fix: Add `user_month_stats`, `admin_month_stats`
