+++ src/ai/backend/manager/services/resource/actions/admin_month_stats.py
@@ -1,0 +1,39 @@
+from dataclasses import dataclass
+from typing import Any, Optional, Self, override
+
+from ai.backend.manager.actions.action import BaseActionResult
+from ai.backend.manager.services.resource.base import ResourceAction
+
+
+@dataclass
+class AdminMonthStatsAction(ResourceAction):
+    @override
+    def entity_id(self) -> str:
+        # TODO: ?
+        return ""
+
+    @override
+    def operation_type(self):
+        return "admin_month_stats"
+
+
+@dataclass
+class AdminMonthStatsActionResult(BaseActionResult):
+    stats: list[Any]
+
+    @override
+    def entity_id(self) -> Optional[str]:
+        return None
+
+    @override
+    def status(self) -> str:
+        return "success"
+
+    @override
+    def description(self) -> Optional[str]:
+        return ""
+
+    def __eq__(self, other: Any) -> bool:
+        if not isinstance(other, Self):
+            return False
+        return True
