+++ src/ai/backend/manager/services/resource/processors.py
@@ -1,4 +1,8 @@
 from ai.backend.manager.actions.processor import ActionProcessor
+from ai.backend.manager.services.resource.actions.admin_month_stats import (
+    AdminMonthStatsAction,
+    AdminMonthStatsActionResult,
+)
 from ai.backend.manager.services.resource.actions.check_presets import (
     CheckResourcePresetsAction,
     CheckResourcePresetsActionResult,
@@ -15,7 +19,14 @@ from ai.backend.manager.services.resource.actions.usage_per_month import (
     UsagePerMonthAction,
     UsagePerMonthActionResult,
 )
-from ai.backend.manager.services.resource.actions.usage_per_period import UsagePerPeriodAction, UsagePerPeriodActionResult
+from ai.backend.manager.services.resource.actions.usage_per_period import (
+    UsagePerPeriodAction,
+    UsagePerPeriodActionResult,
+)
+from ai.backend.manager.services.resource.actions.user_month_stats import (
+    UserMonthStatsAction,
+    UserMonthStatsActionResult,
+)
 from ai.backend.manager.services.resource.service import ResourceService
 
 
@@ -25,6 +36,8 @@ class ResourceProcessors:
     recalculate_usage: ActionProcessor[RecalculateUsageAction, RecalculateUsageActionResult]
     usage_per_month: ActionProcessor[UsagePerMonthAction, UsagePerMonthActionResult]
     usage_per_period: ActionProcessor[UsagePerPeriodAction, UsagePerPeriodActionResult]
+    user_month_stats: ActionProcessor[UserMonthStatsAction, UserMonthStatsActionResult]
+    admin_month_stats: ActionProcessor[AdminMonthStatsAction, AdminMonthStatsActionResult]
 
     def __init__(self, service: ResourceService) -> None:
         self.list_presets = ActionProcessor(service.list_presets)
@@ -32,3 +45,5 @@ class ResourceProcessors:
         self.recalculate_usage = ActionProcessor(service.recalculate_usage)
         self.usage_per_month = ActionProcessor(service.usage_per_month)
         self.usage_per_period = ActionProcessor(service.usage_per_period)
+        self.user_month_stats = ActionProcessor(service.user_month_stats)
+        self.admin_month_stats = ActionProcessor(service.admin_month_stats)
